# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Discord Clone API
  version: 1.0.0
  description: |
    API REST pour le clone Discord.
    
    ## Authentification
    Toutes les routes (sauf `/auth/register` et `/auth/login`) nécessitent un token JWT.
    
    ```
    Authorization: Bearer <access_token>
    ```
  contact:
    name: Discord Clone Team
    email: team@discord-clone.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development
  - url: https://api.discord-clone.example.com/api/v1
    description: Production
...
paths:
...
  /servers/{serverId}/invites:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    get:
      tags: [Invites]
      summary: List server invites
      operationId: listInvites
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of invites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Invites]
      summary: Create invite
      operationId: createInvite
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteWithUrl'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /invites/{code}:
    parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          example: abc123XY
    get:
      tags: [Invites]
      summary: Get invite info (public)
      operationId: getInvite
      responses:
        '200':
          description: Invite info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteInfo'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
    delete:
      tags: [Invites]
      summary: Delete invite
      operationId: deleteInvite
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Invite deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ServerId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Gone:
      description: Gone (expired or used)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    Invite:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        server_id:
          type: string
          format: uuid
        created_by:
          type: object
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
        uses:
          type: integer
        max_uses:
          type: integer
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    InviteWithUrl:
      allOf:
        - $ref: '#/components/schemas/Invite'
        - type: object
          properties:
            invite_url:
              type: string
              format: uri
    InviteInfo:
      type: object
      properties:
        data:
          type: object
          properties:
            code:
              type: string
            server:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
                icon_url:
                  type: string
                  nullable: true
                member_count:
                  type: integer
            created_by:
              type: object
              properties:
                username:
                  type: string
                avatar_url:
                  type: string
                  nullable: true
            expires_at:
              type: string
              format: date-time
              nullable: true
    InviteList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Invite'
    CreateInviteRequest:
      type: object
      properties:
        max_uses:
          type: integer
          minimum: 1
          nullable: true
        expires_in_hours:
          type: integer
          minimum: 1
          nullable: true
...